fixed most gcc warnings

+++ ktx/src/combat.c
@@ -20,7 +20,7 @@
  *  Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  *
  *
- *  $Id: combat.c,v 1.23 2006/05/23 01:17:47 ult_ Exp $
+ *  $Id: combat.c,v 1.24 2006/05/25 04:48:48 ult_ Exp $
  */
 
 #include "g_local.h"
@@ -178,7 +178,7 @@ void T_Damage( gedict_t * targ, gedict_t * inflictor, gedict_t * attacker, float
 	if ( !targ->s.v.takedamage || ISDEAD( targ ) )
 		return;
 
-	if ( cvar("k_midair") )
+	if ( (int)cvar("k_midair") )
 		midair = true;
 
 	wp_num = attacker->s.v.weapon;
@@ -421,7 +421,7 @@ void T_Damage( gedict_t * targ, gedict_t * inflictor, gedict_t * attacker, float
 			{
 				attacker->ps.dmg_g += dmg_dealt;
 				targ->ps.dmg_t += dmg_dealt;
-				if ( cvar("k_dmgfrags") )
+				if ( (int)cvar("k_dmgfrags") )
 				{
 					dmg_frags = attacker->ps.dmg_g / 100;
 					attacker->s.v.frags = attacker->s.v.frags - attacker->ps.dmg_frags + dmg_frags;
